+++ elements/ipv6/LookupIP6Route_kernel.cu
@@ -144,7 +144,7 @@ __device__ uint64_t ntohll(uint64_t val)
 
 __global__ void ipv6_route_lookup_cuda(
         struct datablock_kernel_arg **datablocks,
-        uint32_t count, uint16_t *batch_ids, uint16_t *item_ids,
+        uint32_t count, uint8_t *batch_ids, uint16_t *item_ids,
         uint8_t *checkbits_d,
         Item** __restrict__ tables_d,
         size_t* __restrict__ table_sizes_d)
@@ -152,7 +152,7 @@ __global__ void ipv6_route_lookup_cuda(
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (idx < count) {
-        uint16_t batch_idx = batch_ids[idx];
+        uint8_t batch_idx = batch_ids[idx];
         uint16_t item_idx  = item_ids[idx];
         struct datablock_kernel_arg *db_dest_addrs = datablocks[dbid_ipv6_dest_addrs_d];
         struct datablock_kernel_arg *db_results    = datablocks[dbid_ipv6_lookup_results_d];
